rowblk: move unsafe_test.go from sstable
The unit tests in unsafe_test.go were vestiges from when the rowblk
implementation was in the sstable package. The sstable package versions of
these functions (getBytes and decodeVarint) were unused outside of tests. This
commit moves these unit tests into the rowblk package where they exercise the
versions of these functions that are actually in use, and removes the unused
sstable package funcs.
